Overhaul Component type ids
Type ids are now generated by the ComponentFactory. The factory
does not directly use the ids, but since it's the only valid way
to acquire an id, it should serve as a reminder to properly
register a new component class.

In addition, the TYPE_ID field of each component class is now
a constant data field instead of a function. This makes sense
now because the static ids (of the C++ components) are generated
at DLL load time instead of just-in-time when the typeId()
member function is called. This change also makes possible a
more consistent interface for Lua defined component classes.