Explicit init and shutdown for ScriptEntityFilter
This fixes a crash during the destruction of the Engine object.
The engine's EntityManager is destroyed first. During
destruction of the Lua state, any ScriptEntityFilters try to
unregister themselves from the (already destroyed) entity
manager.

To prevent this, we shift the blame on the users (script
authors), requiring them to explicitly initialize and shutdown
the filter, similar to the C++ filters.