Added partial code so new GUI works with tutorial
The code I've added stops the tutorial from breaking immediately with the new assets, but it breaks part way through when it comes to call the compound list. Since I've replaced the listbox with a set of windows to represent compounds, I'm not sure how the code will need to be adjusted and leave it in the hands of someone more capable than me.