0.3.3 release bugfixing (#492)
Fixed a bug where the game would crash if a microbe died while having a
split organelle.
Reverting flagellum momentum back to 12.5.
Fixed a crash where if an organelle was completely surrounded and tried
to split, no space would be found.
Changed the algorithm to one similar to the one used in
procedural_microbes.lua, but i think moving adyacent organelles
recursively would be a more realistic implementation for later.
The pause function was unimplemented. Changing it to close fixed the
issue.
Fixed a bug where a microbe could sometimes reproduce without having to
split all of it's organelles.
Now the microbes will poop outside of themselves, instead of reabsorbing
all of the ejected compounds.
Update script to copy .ps and .vs GLSL files into dist folder.