Fixed bug venting everything
Current bugs:
3 cells spawn after dividing when tutorial is completed and after that
(but not when tutorial is skipped)

If tutorial is complete and the game is started again from the main menu
and the tutorial is skipped the game crashes. Also other way around
works: if you first play normally and then play the tutorial the game
crashes when exiting the editor.