895 pilus sound (#923)
* Decreased pilus sound

And also put a sound modifier variable into the playSoundWithDistance function so that changing the volumes of different sounds is much simpler in the future. This could also come in handy when a difference in volume of the same audio clip at different scenarios is required.

* Just changed the sound file

Previous commit was not very optimal, so I just replaced the audio file of the pilus hit sound effect with a quieter version (credits to DPh Kraken), and reverted the previous commit's changes.