Implemented a basic limiting on the number of things that can be (#1361)
spawned per frame, this does slightly alter the spawn system behaviour
when respawning to not spawn a huge number of cells. Also this does
still cause some slowdowns as the things are quickly spawned over a
few frames, but that can't really be fixed with the current spawn
system that every now and then runs and wants to spawn a lot of
stuff. We should really rewrite the spawn system to be more time based
instead of checking has the player moved enough every now and then and
spawning stuff