Added line graph UI for the report tab statistics (#1795)
* Added basic line graph, still needs major work

* Added chart legends

* Improved the chart legend a bit.
Moved line chart files into its own folder. Added a default icon
for the icon legend (a blank white circle). Set focus mode to None
on some of editor buttons and Click on the species name line edit.

* Fixed csproj

* Merged LineChartDrawer into LineChart class

* Improved the dropdown legend a bit by adding icons

* Fixed dropdown icons drawn outside of the popupmenu area

* Removed redundant name qualifier

* Tried to fix formatting second try

* Tried to fix formatting third try

* Made the lines not offset from the chart origin

Styled the axes line, Fixed ClearTooltips() method, renamed ChartPoint
to DataPoint and added some special handling for data point markers
that have to be drawn at zero values (so it doesn't overlap axes line).

* Fixed up a few formatting issues

* Removed unused tooltip group nodes

* Made the chart drawing happen in the Drawer node.

To make the renderings position more in sync inside the Drawer area

* Fixed cytoplasm organelle selection button having different font from the rest

* Renamed 'TooltipHandler', tweaked double membrane tooltip

* Added PLACEHOLDER tags to the axes label, and ran update_localization.rb

* Fixed up some texts for localization.

Removed '100 Myr' text on the time indicator, and added AUTO_EVO key for
the auto-evo button on the report tab (currently hidden).

* Forgot to commit this

* Ran update_localization.rb

* Fixed membrane tab color picker font size.

Its theme was unset for some reason, so added back a theme to the color
picker control to lower the font size to how it was before.

* Let godot fix the theme id numbering

* Did the suggested changes

* Fixed silly mistakes

* Fixed locale string and reword some comments

* Cleared AUTO_EVO string in other locales