Fixed rigidity slider not unlocking after undo
Added a method to update the state of the rigidity slider every time mutation points is updated.

However; not sure if this was the case before but if the slider is locked due to dragging, when undoing it still thinks the mouse is grabbed on the slider so it will immediately move towards the mouse whenever crossed over it.