Add an empty DynamicFont to compound and environment panels' expand/collapse buttons
The new Jura font version seems to have a bigger height than the previous,
this causes these buttons to be slightly taller (due to the built-in label
nodes inside of them). An empty DynamicFont is set as the custom font
overrides to nullify their label nodes height and making the buttons
symmetrical again.