Squash all commits of add cost factor to editor
Adds a constant factor affecting the MP cost of all cell editor actions. In normal microbe stage this is 1, so there is no effect. In the multicelluar editor, the value is 0.5 (this is a simulation parameter constant and can be changed). Values are multiplied and cast to (int) so that a player can always perform 2 actions where they would have been able to perform one, since switching to fractional MP costs sounded annoying (and we might need to handle precision issues since floats seem to be the paradigm).