Handle sub-ingested objects properly and fixed enzymes functionality.
Objects that are engulfed by another object that you currently engulf
now transfer its "ownership" of those objects to you, allowing you to
absorb the nutrients of what they've ingested. This "should" also fix
physics isses.

Enzymes functionality somehow broke, which I've now fixed, not very
confident with the overall design but it's what it is for now.
Fixed many other things that are previously overlooked.