Created a new concept of customized registry types to make difficulty and auto-evo settings saving better
basically I ended up setting up a system where a customized version of a registry type can be created that can be edited freely while making saving and loading work both with the customized and default instances.

This also now makes it so that we can update difficulty related numbers and old saves will benefit.

As far as I could test this doesn't seem to introduce any save compatibility problems