Fix current light level not working correctly
Fixing slightly flawed approach in #4276.

Moved `DayNightCycle` instance out of `CreatureStageBase` higher to `GameWorld` so that it's accessible to more logic that needs it. Also refactored some TODO comments regarding updating average sunlight in patches.

Editor light level is now changed to be a daylight fraction instead. This is to make it easier to apply the correct editor-specific time of day to all patches in the patch map.