Modify custom pop ups to adjust to changing heights (#4523)
* Make custom pop up menu height adjust to change

Before it was basically just setting size to the predetermined custom
minimum size every time it was opened. Now it calculates size based off
of the container of it's elements and changes the height to encompass
the height of all the nodes.

* Make CalculateSize consider title height

In the base method it did not take into account the empty space
separating multiple lines of text. This derivative method takes into
account the title height changing.

* Modify OrganellePopupMenu to have proper size

The extended margins that existed previously did not work with the
change to CustomPopup sizing and caused unintended behaviour.

* Remove debug print statement

* Fix typo in comment

Co-authored-by: Henri Hyyryläinen <henri.hyyrylainen@gmail.com>

* Improve comments

* Minimize line breaks in code

* Get label font first

Gets the label font first, then if not available back up to the normal
font.

* Move workaround to protected function

Reduces the exposure of variables from private to protected that don't
need to be.

* Add comment explaining size calculation logic

* Move clip control init to node path export

* Fix workaround method summary comment formatting

* Fix clip control nodepath

Added disposing logic and removed nullable mark from nodepath

* Restructure CustomPopupMenu to use PanelContainer

* Remove old code