Fixed the big iron collision model being way off (#4797)
by removing a rotation set in the Godot editor for it. Also made the automatic scale up for physics to be multiplied by 100 instead of 90 to get it now to match the visual model. Small iron still has this bug. Also fixed a typo I noticed elsewhere in the code.