Fix Issue #5230
Issue #5230: Images in the Thriveopedia are cut off before displaying the full image, as shown below.

https://ibb.co/yY5pz3V

This should fix issue #5230.

Changed the stretch mode in the 'Icon' image to be Keep Aspect Ratio. This keeps it within the box.

Changed the stretch mode in the 'Model' image to be Keep Aspect Ratio - it was having a similar problem.

Screenshot of fixes:

https://ibb.co/0DdybDq