Tweaked Silica, Calcium Carbonate, Chitin, Cellulose
- reduced silica osmoregulation effect to make cellulose more unique, and to nerf silica a bit. Also slightly increased resource intake amount.
- slightly reduced toxin resistance of chitin
- made calcium carbonate more vulnerable to physical resistance. Made it have less of a bonus to osmoregulation, and more ability to absorb compounds. 

This makes it so that calcium carbonate is more special in its ability to serve as the ultimate form of toxin resistance, but is not totally invulnerable. 

Change to calcium carbonate also makes silica more unique. Silica serves as the ultimate "fortress" strategy, completely shutting yourself off to the world in exchange for supreme protection.