Partially fixed the timeline tab having scrollbars unnecessarily
when the content wasn't tall enough to warrant them. Still sometimes causes the scrollbars to appear on the local events list when unnecessary, but this makes that less likely. That seems to be kind of related to the global list and local sizes interacting weirdly with the container. Swapping between global and local events fixes the too long scrollbar.