Slightly Buff Aerobic Respiration
Nerfing aerobic respiration and the rate of oxygen accumulation was a double hit which made it really hard to justify oxygen for the first 20 turns. With this balancing, by turn 20 or slightly earlier, oxygen should emerge as the most efficient form of respiration, and should continue to become more powerful by the end of the stage.