Tweaks to oxygen generation, respiration output, iron metabolism nerf (#6358)
* Balancing changes related to metabolism

* Corrected UTF-8 hopefully

* Slightly Decrease Cytoplasm ATP Output

Makes it so that parts besides your cytoplasm are more important early on, making metabolisms more characteristic of themselves in smaller cells. Otherwise, iron doesn't really behave as it should since it won't used as much to create ATP with enough being generated from cytoplasm.

* Reduced Iron Input Amount Needed To Reduce Spawn Death Unfairness

If players dominantly rely on iron for ATP generation, they could sometimes start dying immediately upon death because it can take a bit to find iron.

* Reduced Iron Input More

I think it's now fairly balanced where players are constrained by iron's characteristics, but don't die if they aren't a few seconds away from iron. Iron is now really benefited from storage and chemoreceptors

* Slightly Speed Up Rate of Oxygen Accumulation

Prior pace was pretty slow and could halter at times. This should make it so that once oxygen starts building, it'll continue to build up unless something remarkable happens

* Slightly Buff Aerobic Respiration

Nerfing aerobic respiration and the rate of oxygen accumulation was a double hit which made it really hard to justify oxygen for the first 20 turns. With this balancing, by turn 20 or slightly earlier, oxygen should emerge as the most efficient form of respiration, and should continue to become more powerful by the end of the stage.

* Thermosynthesis Now Creates Glucose

Akin to Rathalos' suggestion, thermosynthesis now creates glucose. It is balanced now to be a solid source of glucose when more simple/small, but to be rather troublesome to effectively scale up.

* Undid the thermosynthesis to glucose conversion for now

as that was more not-settled change of the balance thing yet

* Try to give slightly better error message in native fetch

* Fix error output message a bit

* Try buffering entire file in memory if native download fails

* Added safety checks to native lib upload tool

* Check download size before decoding

---------

Co-authored-by: Deus-Codes <yn0048a@student.american.edu>
Co-authored-by: Deus-Codes <103910570+Deus-Codes@users.noreply.github.com>