Disabled focusing on auto-evo prediction text popup
so that a white outline that goes outside the popup box is not drawn. For some reason the screen reader accessibility focus was triggering this, I'm not sure what else could be done besides disabling it entirely on this one node. Also Godot did a lot of random changes that I think are all fully correct.