Remade MP calculation system (again) (#6460)
* Started on readonly interfaces for species data comparison feature

* Renamed colour in species for more clarity and conflict resolve

when interfaces are both a cell type and a species

* Adjusted interfaces and added new ones for readonly

access to various microbe species data

* Removed include that is no longer needed thanks to our list helper

new method

* One more interface

* Initial implementation of the species comparer

* Created a species facade for applying edits on top of a species

for use with the species comparer

* Switched the microbe editor to use the new MP calculation

partially, not yet used for MP prediction or the context menu. And not implemented for later stage editors yet.

* Mark PauseMenu as singleton

* Made editor MP preview use the new system as well

* Made placing stuff on top of cytoplasm not incur the remove cost

that would normally apply to the cytoplasm

* Made facade capture initial positions

and added object caching as it needs to create proxy objects for all edits due to needing to capture the initial position

* Added a test to ensure archive object types are unique

* Fixed bug with resolved moves not being removed

* Added support for some things in facade for tests

* Started on the interfaces for multicell and macroscopic

new MP calculation approach

* The new interfaces to go with the previous commit

* Started converting the MP tests to the new calculation

type based on the facade rather than the editor history

* Dummy files for macroscopic and multicellular MP

so that they won't crash

* Converted one more test

* Fixed endosymbiont actions not being free

and updated a few more MP tests. Needed to add new endosymbiont property tracking and save properties, which hopefully didn't break anything.

* Updated a bunch more MP tests

* Fully fixed all MP tests

* Refresh facade if reading the enumerator first

* Added old position to organelle upgrade action

as the facade class may need to rely on it in some edit situations

* Refactored subsequent facade actions applying

to search the already applied organelles list first before looking at old data, this simplified the logic quite a bit but required position info for upgrade action

* Cleaned up some code and made a good way to disable localization in tests

or at least a simpler way than just hunting for what to disable

* Comment out multicellular and macroscopic edits

so that there aren't any warnings as the microbe part should now be fully done

* Suppress some engine-originating error popups

as we don't want players specifically to report these

* Fixed some style problems

* A few more style fixes

* A few more automatic formatting changes