Fixed the industrial stage (pause menu duplicate instance) (#6516)
* Fixed industrial stage accidentally still having a pause menu

instead of using the global singleton

* Made pause menu instance work for directly started scenes

as otherwise they would run into an error as the pause menu was not available so soon for use. Hopefully no one tries to use too complex operations on the pause menu before _Ready now gets called on it. Also unregister the pause menu on tree exit, probably not needed but might be nice for things to be symmetric.