5243 positioning controller tab movement prompts closer to tabs (#6579)
* Shrunk min size of TabButtons

The minimum size of TabButtons was set to 1200 in the options menu. As the button container child was set to expand, this caused the tab prompts to be pushed out to the edge. This changes that custom minimum to something that places the prompts close to the buttons.

* Tab button size now accounts for other languages

* Added SetCustomMinimumSize to TabButtons

Added the function SetCustomMinimumSize to TabButtons, which when called will calculate an appropriate custom minimum size so that all of the buttons and the left and right indicators will fit on a single line. Also added calling that new function in OptionsMenu when applicable.

* Corrected formatting errors

I forgot to run the formatting checks before I pushed, those have now been corrected.

* Removed duplicate code and swaped to custom invoke

Moved the SetCustomMinimumSize calls into the more generic OnTranslationChanged. It also now uses Thrive's custom invoker instead of the generic Godot one.

* Used disposable string name in TabButtons

* Fixing formatting error created in merge

* Renamed a variable to a simpler name

---------

Co-authored-by: Henri Hyyryläinen <henri.hyyrylainen@gmail.com>